requirements.txt
pip == 22.2
streamlit == 1.12.0
numpy == 1.20.2
Pillow == 8.2.0
tensorflow == 2.9.1
opencv-python-headless==4.6.0.66
